refactor: TaskScheduler统一纳管周期性任务 + clippy零error + 存储原子接口
- TaskScheduler纳管所有36个周期性任务，tick从500ms降至100ms
- 建立[ALLOWED-INTERVAL]/[ALLOWED-SLEEP]白名单标记规范
- 修复全部clippy错误（pdc自身零error）
- 6个存储模块添加save_dirty原子接口
- 建立硬编码白名单机制[ALLOWED-HARDCODED]，290处历史遗留标记
- 合规检查26项全部通过（0 FAIL，6 WARN为历史遗留）

+++ src/federation/merkle.rs
@@ -4,11 +4,9 @@
 //! 将 key 空间按 shard_count 分片，每片维护独立的根哈希。
 
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 
 use parking_lot::RwLock;
 use rustc_hash::{FxHashMap, FxHashSet};
-use serde::Serialize;
 
 use crate::federation::protocol::MerkleDigestMessage;
 
@@ -47,7 +45,7 @@ impl MerkleTree {
     fn shard_for_key(&self, key: &[u8]) -> u16 {
         let hash = blake3::hash(key);
         let bytes = hash.as_bytes();
-        (u16::from_le_bytes([bytes[0], bytes[1]]) % self.shard_count as u16)
+        u16::from_le_bytes([bytes[0], bytes[1]]) % self.shard_count
     }
 
     /// 重算指定分片的根哈希
@@ -78,9 +76,7 @@ impl MerkleTree {
     /// 更新/插入条目
     pub fn update(&self, key: &[u8], data: &[u8]) {
         let shard = self.shard_for_key(key);
-        self.entries
-            .write()
-            .insert(key.to_vec(), data.to_vec());
+        self.entries.write().insert(key.to_vec(), data.to_vec());
         // 收/发全量同步期间跳过单条重算，结束后由 rebuild_all 一次性重算
         if !self.full_sync_active() {
             self.recompute_shard(shard);
